+++ src/Repositories/CachingRepository.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Saritasa\Repositories;
+namespace Saritasa\LaravelRepositories\Repositories;
 
 use Closure;
 use Illuminate\Contracts\Cache\Repository as CacheRepository;
@@ -10,13 +10,13 @@ use Illuminate\Support\Collection;
 use Saritasa\DingoApi\Paging\CursorRequest;
 use Saritasa\DingoApi\Paging\CursorResult;
 use Saritasa\DingoApi\Paging\PagingInfo;
-use Saritasa\DTO\SortOptions;
-use Saritasa\Exceptions\ModelNotFoundException;
-use Saritasa\Exceptions\RepositoryException;
-use Saritasa\Contracts\IRepository;
+use Saritasa\LaravelRepositories\DTO\SortOptions;
+use Saritasa\LaravelRepositories\Exceptions\ModelNotFoundException;
+use Saritasa\LaravelRepositories\Exceptions\RepositoryException;
+use Saritasa\LaravelRepositories\Contracts\IRepository;
 
 /**
- *
+ * Wrapper for repositories to store data into cache.
  */
 class CachingRepository implements IRepository
 {
@@ -25,10 +25,10 @@ class CachingRepository implements IRepository
      *
      * @var IRepository
      */
-    private $repo;
+    protected $repository;
 
     /**
-     * Cache repository.
+     *  Cache storage implementation.
      *
      * @var CacheRepository
      */
@@ -39,31 +39,45 @@ class CachingRepository implements IRepository
      *
      * @var string
      */
-    private $prefix;
+    protected $prefix;
 
     /**
      * Cache timeout
      *
      * @var int
      */
-    private $cacheTimeout;
+    protected $cacheTimeout;
 
-    /* @var string */
-    private $modelClass;
-
+    /**
+     * Wrapper for repositories to store data into cache.
+     *
+     * @param IRepository $repository Repositories which call will be cached
+     * @param CacheRepository $cacheRepository Cache storage implementation
+     * @param string $prefix Cache prefix
+     * @param int $cacheTimeout Time while cache data will bea actual
+     */
     public function __construct(
         IRepository $repository,
         CacheRepository $cacheRepository,
         string $prefix,
         int $cacheTimeout = 10
     ) {
-        $this->repo = $repository;
+        $this->repository = $repository;
         $this->prefix = $prefix;
         $this->cacheTimeout = $cacheTimeout;
-        $this->modelClass = $repository->getModelClass();
+        $this->cacheRepository = $cacheRepository;
     }
 
-    private function cached(string $key, Closure $dataGetter)
+    /**
+     * Get data from cache.
+     * If it not exists in cache got it in repository and store in cache.
+     *
+     * @param string $key Key to find data in cache
+     * @param Closure $dataGetter Function to get data in original repository
+     *
+     * @return mixed
+     */
+    protected function cached(string $key, Closure $dataGetter)
     {
         if ($this->cacheRepository->has($key)) {
             return $this->cacheRepository->get($key);
@@ -73,94 +87,145 @@ class CachingRepository implements IRepository
         return $result;
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function getModelValidationRules(): array
     {
-        return $this->repo->getModelValidationRules();
+        return $this->repository->getModelValidationRules();
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function findOrFail($id): Model
     {
         $result = $this->cachedFind($id);
         if (!$result) {
-            throw new ModelNotFoundException($this, "$this->modelClass with ID=$id was not found");
+            throw new ModelNotFoundException($this, "{$this->repository->getModelClass()} with ID=$id was not found");
         }
         return $result;
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function findWhere(array $fieldValues): ?Model
     {
         $key = $this->prefix . ":find:" . md5(serialize($fieldValues));
         return $this->cached($key, function () use ($fieldValues) {
-            return $this->repo->findWhere($fieldValues);
+            return $this->repository->findWhere($fieldValues);
         });
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function create(Model $model): Model
     {
-        return $this->repo->create($model);
+        return $this->repository->create($model);
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function save(Model $model): Model
     {
-        $result = $this->repo->save($model);
+        $result = $this->repository->save($model);
         $this->invalidate($model);
         return $result;
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function delete(Model $model): void
     {
-        $this->repo->delete($model);
+        $this->repository->delete($model);
         $this->invalidate($model);
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function get(): Collection
     {
         return $this->cached("$this->prefix:all", function () {
-            return $this->repo->get();
+            return $this->repository->get();
         });
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function getWhere(array $fieldValues): Collection
     {
         $key = $this->prefix . ":get:" . md5(serialize($fieldValues));
         return $this->cached($key, function () use ($fieldValues) {
-            return $this->repo->getWhere($fieldValues);
+            return $this->repository->getWhere($fieldValues);
         });
     }
 
-    public function getPage(PagingInfo $paging, array $fieldValues = null): LengthAwarePaginator
+    /**
+     * {@inheritdoc}
+     */
+    public function getPage(PagingInfo $paging, array $fieldValues = []): LengthAwarePaginator
     {
         $key = $this->prefix . ":page:" . md5(serialize($paging->toArray()) . serialize($fieldValues));
         return $this->cached($key, function () use ($paging, $fieldValues) {
-            return $this->repo->getPage($paging, $fieldValues);
+            return $this->repository->getPage($paging, $fieldValues);
         });
     }
 
-    public function getCursorPage(CursorRequest $cursor, array $fieldValues = null): CursorResult
+    /**
+     * {@inheritdoc}
+     */
+    public function getCursorPage(CursorRequest $cursor, array $fieldValues = []): CursorResult
     {
         $key = $this->prefix . ":page:" . md5(serialize($cursor->toArray()) . serialize($fieldValues));
         return $this->cached($key, function () use ($cursor, $fieldValues) {
-            return $this->repo->getCursorPage($cursor, $fieldValues);
+            return $this->repository->getCursorPage($cursor, $fieldValues);
         });
     }
 
+    /**
+     * Find model in original repository.
+     *
+     * @param string|int id Id to find
+     *
+     * @return Model|null
+     */
     private function find($id): ?Model
     {
         try {
-            return $this->repo->findOrFail($id);
+            return $this->repository->findOrFail($id);
         } catch (ModelNotFoundException $exception) {
             return null;
         }
     }
 
-    private function cachedFind($id) //: Model
+    /**
+     * Find model in cache by id.
+     *
+     * @param string|int $id Id to find
+     *
+     * @return Model|null
+     */
+    protected function cachedFind($id): ?Model
     {
         return $this->cached("$this->prefix:$id", function () use ($id) {
             return $this->find($id);
         });
     }
 
-    private function invalidate(Model $model)
+    /**
+     * Invalidate model in cache.
+     *
+     * @param Model $model Model to invalidate
+     *
+     * @return void
+     */
+    protected function invalidate(Model $model): void
     {
         $key = $this->prefix . ":" . $model->getKey();
         if ($this->cacheRepository->has($key)) {
@@ -169,36 +234,73 @@ class CachingRepository implements IRepository
         $this->cacheRepository->forget("$this->prefix.:all");
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function getModelClass(): string
     {
-        return $this->modelClass;
+        return $this->repository->getModelClass();
     }
 
-    public function __call($name, $arguments)
+    /**
+     * Proxy for get methods.
+     *
+     * @param string $name Method name
+     * @param array $arguments Method arguments
+     *
+     * @return mixed
+     *
+     * @throws RepositoryException
+     */
+    public function __call(string $name, array $arguments)
     {
         if (0 === strpos($name, 'get')) {
             $argHash = md5(serialize($arguments));
             $key = "$this->prefix:$name:$argHash";
             return $this->cached($key, function () use ($name, $arguments) {
-                return call_user_func_array([$this->repo, $name], $arguments);
+                return call_user_func_array([$this->repository, $name], $arguments);
             });
         }
         throw new RepositoryException($this, "Caching repository proxies only get* methods");
     }
 
-    public function __get($name)
+    /**
+     * Proxy for repository params.
+     *
+     * @param string $name Parameter name
+     *
+     * @return mixed
+     */
+    public function __get(string $name)
     {
         return $this->cached("$this->prefix:$name", function () use ($name) {
-            return $this->repo->$name;
+            return $this->repository->$name;
         });
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function getWith(
         array $with,
-        array $withCounts = null,
-        array $where = null,
-        SortOptions $sortOptions = null
+        array $withCounts = [],
+        array $where = [],
+        ?SortOptions $sortOptions = null
     ): Collection {
-        // TODO: Implement getWith() method.
+        $key = $this->prefix . ":get:" . md5(serialize($with) . serialize($withCounts) . serialize($where));
+        return $this->cached($key, function () use ($with, $withCounts, $where, $sortOptions) {
+            return $this->repository->getWith($with, $withCounts, $where, $sortOptions);
+        });
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function count(): int
+    {
+        $key = $this->prefix . ":all:count";
+        return $this->cached($key, function () {
+            return $this->repository->count();
+        });
     }
 }