+++ tests/JoinRelationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Saritasa\Repositories\Tests;
+namespace Saritasa\LaravelRepositories\Tests;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\RelationNotFoundException;
@@ -11,14 +11,15 @@ use Illuminate\Database\Eloquent\Relations\HasOne;
 use Illuminate\Database\Eloquent\Relations\MorphMany;
 use PHPUnit\Framework\TestCase;
 use Saritasa\Exceptions\NotImplementedException;
-use Saritasa\Repositories\EloquentRepository;
+use Saritasa\LaravelRepositories\Exceptions\RepositoryException;
+use Saritasa\LaravelRepositories\Repositories\Repository;
 
 /**
  * Check join relation repository method.
  */
 class JoinRelationTest extends TestCase
 {
-    /** @var EntitiesRepository */
+    /** @var EntityRepository */
     private $repository;
 
     /**
@@ -28,7 +29,7 @@ class JoinRelationTest extends TestCase
      * @param array $data
      * @param string $dataName
      *
-     * @throws \Saritasa\Exceptions\RepositoryException
+     * @throws RepositoryException
      */
     public function __construct(string $name = null, array $data = [], string $dataName = '')
     {
@@ -251,7 +252,7 @@ class Note extends Model
 /**
  * Fake user records repository class. Has debug methods for performing tests.
  */
-class EntitiesRepository extends EloquentRepository
+class EntitiesRepository extends Repository
 {
     /**
      * Fake user records repository class. Has debug methods for performing tests.